% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clinicalTables}
\alias{clinicalTables}
\title{Tables in the cdm_reference that contain clinical information}
\usage{
clinicalTables()
}
\value{
A character vector with table names.
}
\description{
This function provides a list of allowed inputs for the \code{omopTableName}
argument in \code{summariseClinicalRecords()}.
}
\examples{
library(OmopSketch)

clinicalTables()

}
