% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{dateRange-startDate}
\alias{dateRange-startDate}
\title{Helper for consistent documentation of \code{dateRange}.}
\arguments{
\item{dateRange}{A vector of two dates defining the desired study period.
Only the \code{start_date} column of the OMOP table is checked to ensure it falls within this range.
If \code{dateRange} is \code{NULL}, no restriction is applied.}
}
\description{
Helper for consistent documentation of \code{dateRange}.
}
\keyword{internal}
