% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{interval}
\alias{interval}
\title{Helper for consistent documentation of \code{interval}.}
\arguments{
\item{interval}{Time interval to stratify by. It can either be "years", "quarters", "months" or "overall".}
}
\description{
Helper for consistent documentation of \code{interval}.
}
\keyword{internal}
