% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseClinicalRecords.R
\name{missingDataOptions}
\alias{missingDataOptions}
\title{Additional customisable arguments for summarising missingness in the \code{summariseClinicalRecords()} function.}
\usage{
missingDataOptions()
}
\value{
A named list of default options.
}
\description{
This function provides a list of allowed inputs for the \code{.option} argument in
\code{summariseClinicalRecords}, and their corresponding default values.
}
\examples{
missingDataOptions()

}
