% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockOmopSketch.R
\name{mockOmopSketch}
\alias{mockOmopSketch}
\title{It creates a mock database for testing OmopSketch package}
\usage{
mockOmopSketch(con = NULL, writeSchema = NULL, numberIndividuals = 100)
}
\arguments{
\item{con}{A DBI connection to create the cdm mock object. By default, the connection would be "duckdb".}

\item{writeSchema}{Name of an schema on the same connection with writing permissions.}

\item{numberIndividuals}{Number of individuals to create in the cdm reference.}
}
\value{
A mock cdm_reference object created following user's specifications.
}
\description{
It creates a mock database for testing OmopSketch package
}
\examples{
\donttest{
mockOmopSketch(numberIndividuals = 1000,
               writeSchema = NULL,
               con = NULL)
}
}
