% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopSketch-package.R
\docType{package}
\name{OmopSketch-package}
\alias{OmopSketch}
\alias{OmopSketch-package}
\title{OmopSketch: An R Package to Characterise the OMOP tables in a database.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package uses the OMOP tables of the database to summarise key information that may be used to assess its suitability for a certain study. The package offers basic functionalities designed for one-time execution per database release, as well as additional features for ongoing, study-specific analysis. These capabilities enable detailed exploration of the domains and feasibility counts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopSketch/}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/OmopSketch/issues}
}

}
\author{
\strong{Maintainer}: Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})

Authors:
\itemize{
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk} (\href{https://orcid.org/0000-0002-8462-8668}{ORCID})
  \item Elin Rowlands \email{elin.rowlands@ndorms.ox.ac.uk}
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

Other contributors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID}) [contributor]
}

}
\keyword{internal}
