% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{plot-doc}
\alias{plot-doc}
\title{Helper for consistent documentation for plots.}
\arguments{
\item{style}{Visual theme to apply. Character, or \code{NULL}. If a character,
this may be either the name of a built-in style (see
\code{visOmopResults::plotStyle()}), or a path to a .yml file that defines a
custom style. If \code{NULL}, the function will use the explicit default style,
unless a global style option is set (see \code{visOmopResults::setGlobalPlotOptions()})
or a _brand.yml file is present (in that order).}

\item{type}{Character string indicating the output plot format. See
\code{visOmopResults::plotType()} for the list of supported plot types. If
\code{type = NULL}, the function will use the global setting defined via
\code{visOmopResults::setGlobalPlotOptions()} (if available); otherwise, a
standard \code{ggplot2} plot is produced by default.}
}
\description{
Helper for consistent documentation for plots.
}
\keyword{internal}
