% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInObservation.R
\name{plotInObservation}
\alias{plotInObservation}
\title{Create a gt table from a summarised omop_table.}
\usage{
plotInObservation(result, facet = NULL, colour = NULL)
}
\arguments{
\item{result}{A summarised_result object (output of summariseInObservation).}

\item{facet}{Columns to face by. Formula format can be provided. See possible
columns to face by with: \code{visOmopResults::tidyColumns()}.}

\item{colour}{Columns to colour by. See possible columns to colour by with:
\code{visOmopResults::tidyColumns()}.}
}
\value{
A ggplot showing the table counts
}
\description{
Create a gt table from a summarised omop_table.
}
