% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseObservationPeriod.R
\name{plotObservationPeriod}
\alias{plotObservationPeriod}
\title{Create a plot from the output of summariseObservationPeriod().}
\usage{
plotObservationPeriod(
  result,
  variableName = "number subjects",
  plotType = "barplot",
  facet = "cdm_name",
  colour = "observation_period_ordinal"
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{variableName}{The variable to plot it can be: "number subjects",
"records per person", "duration" or "days to next observation period".}

\item{plotType}{The plot type, it can be: "barplor", "boxplot" or
"density".}

\item{facet}{Elements to facet by, it can be "cdm_name",
"observation_period_ordinal", both or none.}

\item{colour}{Elements to color by, it can be "cdm_name",
"observation_period_ordinal", both or none.}
}
\value{
A ggplot2 object.
}
\description{
Create a plot from the output of summariseObservationPeriod().
}
