% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Create a ggplot2 plot from the output of summariseTrend().}
\usage{
plotTrend(result, output = NULL, facet = "type", colour = NULL, style = NULL)
}
\arguments{
\item{result}{A summarised_result object (output of summariseTrend).}

\item{output}{The output to plot.  Accepted values are: \code{"record"}, \code{"person"},
\code{"person-days"}, \code{"age"}, and \code{"sex"}.
If not specified, the function will default to:
\itemize{
\item the only available output if there is just one in the results, or
\item \code{"record"} if multiple outputs are present.
}}

\item{facet}{Columns to face by. Formula format can be provided. See possible
columns to face by with: \code{visOmopResults::tidyColumns()}.}

\item{colour}{Columns to colour by. See possible columns to colour by with:
\code{visOmopResults::tidyColumns()}.}

\item{style}{Which style to apply to the plot, options are: "default", "darwin"
and NULL (default ggplot style). Customised styles can be achieved by modifying the returned ggplot object.}
}
\value{
A ggplot showing the table counts
}
\description{
Create a ggplot2 plot from the output of summariseTrend().
}
\examples{
\donttest{
library(dplyr)
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

result <- summariseTrend(cdm,
  episode = "observation_period",
  output = c("person-days", "record"),
  interval = "years",
  ageGroup = list("<=40" = c(0, 40), ">40" = c(41, Inf)),
  sex = TRUE
)

plotTrend(result, output = "record", colour = "sex", facet = "age_group")


cdmDisconnect(cdm)
}
}
