% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{sample}
\alias{sample}
\title{Helper for consistent documentation of \code{sample}.}
\arguments{
\item{sample}{Either an integer or a character string.
If an integer (n > 0), the function will first sample \code{n} distinct
\code{person_id}s from the \code{person} table and then subset the input tables to
those subjects.
If a character string, it must be the name of a cohort in the \code{cdm}; in
this case, the input tables are subset to subjects (\code{subject_id}) belonging
to that cohort.
Use \code{NULL} to disable subsetting (default value).}
}
\description{
Helper for consistent documentation of \code{sample}.
}
\keyword{internal}
