% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{style}
\alias{style}
\title{Helper for consistent documentation of \code{style}.}
\arguments{
\item{style}{Named list that specifies how to style the different parts of the gt or flextable table generated.
Accepted style entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body.
Alternatively, use "default" to get visOmopResults style, or NULL for gt/flextable style.
Keep in mind that styling code is different for gt and flextable. Additionally, "datatable" and "reactable" have their own style functions.
To see style options for each table type use \code{visOmopResults::tableStyle()}}
}
\description{
Helper for consistent documentation of \code{style}.
}
\keyword{internal}
