% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseEntryCharacteristics.R
\name{summariseEntryCharacteristics}
\alias{summariseEntryCharacteristics}
\title{Summarise characteristics of the individuals at the start of their
observation period.}
\usage{
summariseEntryCharacteristics(cdm, byYear = FALSE)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{byYear}{Whether to stratify the reuslts by year.}
}
\description{
Summarise characteristics of the individuals at the start of their
observation period.
}
