% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseInObservation.R
\name{summariseInObservation}
\alias{summariseInObservation}
\title{Create a summarised result with the number of people in observation during a specific interval of time}
\usage{
summariseInObservation(
  observationPeriod,
  unit = "year",
  unitInterval = 1,
  ageGroup = NULL,
  sex = FALSE
)
}
\arguments{
\item{observationPeriod}{observation_period omop table.}

\item{unit}{Whether to stratify by "year" or by "month".}

\item{unitInterval}{Number of years or months to stratify with.}

\item{ageGroup}{A list of age groups to stratify results by.}

\item{sex}{Whether to stratify by sex (TRUE) or not (FALSE).}
}
\value{
A summarised_result object with the summarised data.
}
\description{
Create a summarised result with the number of people in observation during a specific interval of time
}
