% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseObservationPeriod.R
\name{summariseObservationPeriod}
\alias{summariseObservationPeriod}
\title{Create a gt table from a summarised omop_table.}
\usage{
summariseObservationPeriod(observationPeriod, unit = "year", unitInterval = 1)
}
\arguments{
\item{observationPeriod}{observation_period omop table}

\item{unit}{Whether to stratify by "year" or by "month"}

\item{unitInterval}{Number of years or months to stratify with}
}
\value{
A summarised_result object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
