% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOmopSnapshot.R
\name{summariseOmopSnapshot}
\alias{summariseOmopSnapshot}
\title{Summarise OMOP database and create snapshot}
\usage{
summariseOmopSnapshot(cdm)
}
\arguments{
\item{cdm}{A cdm reference object}
}
\value{
A summarised result object
}
\description{
Summarise OMOP database and create snapshot
}
\examples{
\donttest{
library(dplyr)
library(CDMConnector)
library(DBI)
library(duckdb)
library(OmopSketch)

# Connect to Eunomia database

if (Sys.getenv("EUNOMIA_DATA_FOLDER") == "") Sys.setenv("EUNOMIA_DATA_FOLDER" = tempdir())
if (!dir.exists(Sys.getenv("EUNOMIA_DATA_FOLDER"))) dir.create(Sys.getenv("EUNOMIA_DATA_FOLDER"))
if (!eunomia_is_available()) downloadEunomiaData()

con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())

cdm <- CDMConnector::cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")

# Run OMOP database snapshot

snapshot <- summariseOmopSnapshot(cdm)
}
}
