% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOmopTable.R
\name{summariseOmopTable}
\alias{summariseOmopTable}
\title{Summarise an omop_table from a cdm_reference object. You will obtain
information related to the number of records, number of subjects, whether the
records are in observation, number of present domains and number of present
concepts.}
\usage{
summariseOmopTable(omopTable, byYear = FALSE)
}
\arguments{
\item{omopTable}{An omop_table object.}

\item{byYear}{Whether to stratify the analysis by year.}
}
\description{
Summarise an omop_table from a cdm_reference object. You will obtain
information related to the number of records, number of subjects, whether the
records are in observation, number of present domains and number of present
concepts.
}
