% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOmopTables.R
\name{summariseOmopTables}
\alias{summariseOmopTables}
\title{Summarise the omop_table objects present in a cdm_reference using the
function summariseOmopTable().}
\usage{
summariseOmopTables(cdm, byYear = FALSE)
}
\arguments{
\item{cdm}{An cdm_reference object.}

\item{byYear}{Whether to stratify the analysis by year.}
}
\description{
Summarise the omop_table objects present in a cdm_reference using the
function summariseOmopTable().
}
