% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePerson.R
\name{summarisePerson}
\alias{summarisePerson}
\title{Summarise person table}
\usage{
summarisePerson(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A summarised_result object with the summary of the person table.
}
\description{
Summarise person table
}
\examples{
\donttest{
library(OmopSketch)
library(dplyr, warn.conflicts = FALSE)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

result <- summarisePerson(cdm = cdm)

result |>
  glimpse()

cdmDisconnect(cdm)
}

}
