% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePersonYears.R
\name{summarisePersonDays}
\alias{summarisePersonDays}
\title{Summarise the person years that a cdm contains. You can stratify by ageGroup,
year and sex.}
\usage{
summarisePersonDays(cdm, ageGroup = NULL, byYear = FALSE, bySex = FALSE)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{ageGroup}{A list of age groups to be considered.}

\item{byYear}{Whether to stratify the counts by year.}

\item{bySex}{Whether to stratify the counts by sex.}
}
\value{
A summarised_result object that contains the number of person year
for each of the desired stratifications.
}
\description{
Summarise the person years that a cdm contains. You can stratify by ageGroup,
year and sex.
}
