% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseRecordCounts.R
\name{summariseRecordCounts}
\alias{summariseRecordCounts}
\title{Create a gt table from a summarised omop_table.}
\usage{
summariseRecordCounts(omopTable, unit = "year", unitInterval = 1)
}
\arguments{
\item{omopTable}{A summarised_result object with the output from summariseOmopTable().}

\item{unit}{Whether to stratify by "year" or by "month"}

\item{unitInterval}{Number of years or months to stratify with}
}
\value{
A summarised_result object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
