% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTableQuality.R
\name{summariseTableQuality}
\alias{summariseTableQuality}
\title{Summarise table quality issues in an OMOP table using a specific time interval.}
\usage{
summariseTableQuality(
  cdm,
  omopTableName,
  interval = "overall",
  sex = FALSE,
  ageGroup = NULL,
  sample = NULL,
  dateRange = NULL,
  endBeforeStart = TRUE,
  startBeforeBirth = TRUE
)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}

\item{omopTableName}{A character vector of OMOP table names from the CDM.}

\item{interval}{Time interval to stratify by. It can either be "years", "quarters", "months" or "overall".}

\item{sex}{Logical. Should results be stratified by sex? Default is \code{FALSE}.}

\item{ageGroup}{A list of age groups to stratify results by. Default is \code{NULL}.}

\item{sample}{An integer specifying the number of records to sample from each table.
If \code{NULL}, no sampling is done.}

\item{dateRange}{A vector of two dates defining the desired study period.
Only the \code{start_date} column of the OMOP table is checked to ensure it falls within this range.
If \code{dateRange} is \code{NULL}, no restriction is applied.}

\item{endBeforeStart}{Logical. Whether the function should check for records where the end date
is before the start date. Default is \code{TRUE}.}

\item{startBeforeBirth}{Logical. Whether the function should check for records where the start date
is before the patient's birthdate. Default is \code{TRUE}.}
}
\value{
A \code{summarised_result} object with counts and percentages of potential quality issues,
stratified as specified.
}
\description{
This function assesses potential data quality issues in a given OMOP table, such as records
with an end date before the start date or a start date before the person's birthdate.
}
\examples{
\donttest{
library(dplyr)

cdm <- mockOmopSketch()

qualitySummary <- summariseTableQuality(
  cdm = cdm,
  omopTableName = c("condition_occurrence", "drug_exposure"),
  interval = "years",
  sex = TRUE,
  ageGroup = list("<=20" = c(0, 20), ">20" = c(21, Inf)),
  endBeforeStart = TRUE,
  startBeforeBirth = TRUE
)

qualitySummary |> glimpse()

PatientProfiles::mockDisconnect(cdm = cdm)
}
}
