% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTrend.R
\name{summariseTrend}
\alias{summariseTrend}
\title{Summarise temporal trends in OMOP tables}
\usage{
summariseTrend(
  cdm,
  event = NULL,
  episode = NULL,
  output = "record",
  interval = "overall",
  ageGroup = NULL,
  sex = FALSE,
  inObservation = FALSE,
  dateRange = NULL
)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}

\item{event}{A character vector of OMOP table names to treat as event tables (uses only start date).}

\item{episode}{A character vector of OMOP table names to treat as episode tables (uses start and end date).}

\item{output}{A character vector indicating what to summarise.
Options include \code{"record"} (default), \code{"person"}, \code{"person-days"}, \code{"age"}, \code{"sex"}.
If included, the number of person-days is computed only for episode tables.}

\item{interval}{Time granularity for trends. One of \code{"overall"} (default), \code{"years"},
\code{"quarters"}, or \code{"months"}.}

\item{ageGroup}{A list of age groups to stratify results by.}

\item{sex}{Logical. If \code{TRUE}, stratify results by sex.}

\item{inObservation}{Logical. If \code{TRUE}, the results are stratified to indicate whether each record occurs within an observation period.}

\item{dateRange}{A vector of two dates defining the desired study period.
If \code{dateRange} is \code{NULL}, no restriction is applied.}
}
\value{
A summarised_result object.
}
\description{
This function summarises temporal trends from OMOP CDM tables, considering
only data within the observation period.
It supports both event and episode tables and can report trends such as
number of records, number of subjects, person-days, median age, and number of females.
}
\details{
\itemize{
\item \strong{Event tables}:
Records are included if their \strong{start date} falls within the study period.
Each record contributes to the time interval containing the start date.
\item \strong{Episode tables}:
Records are included if their \strong{start or end date} overlaps with the study
period. Records are \strong{trimmed} to the date range, and contribute to \strong{all}
overlapping time intervals between start and end dates.
}
}
\examples{
\donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

summarisedResult <- summariseTrend(
  cdm = cdm,
  event = c("condition_occurrence", "drug_exposure"),
  episode = "observation_period",
  interval = "years",
  ageGroup = list("<=20" = c(0, 20), ">20" = c(21, Inf)),
  sex = TRUE,
  dateRange = as.Date(c("1950-01-01", "2010-12-31"))
)

summarisedResult |>
  dplyr::glimpse()

cdmDisconnect(cdm = cdm)
}
}
