% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableAllConceptCounts.R
\name{tableAllConceptCounts}
\alias{tableAllConceptCounts}
\title{Create a visual table from a summariseAllConceptCounts() result.}
\usage{
tableAllConceptCounts(result, type = "gt")
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of formatting output table, either "gt" or "flextable".}
}
\value{
A gt or flextable object with the summarised data.
}
\description{
Create a visual table from a summariseAllConceptCounts() result.
}
