% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a gt table from a summarised omop_table.}
\usage{
tableClinicalRecords(summarisedClinicalRecords)
}
\arguments{
\item{summarisedClinicalRecords}{Output from summariseClinicalRecords().}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
\examples{
\donttest{
library(dplyr)
library(CDMConnector)
library(DBI)
library(duckdb)
library(OmopSketch)

# Connect to Eunomia database
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdmFromCon(
 con = con, cdmSchema = "main", writeSchema = "main"
)

# Run summarise clinical tables
summarisedResult <- summariseClinicalRecords(omopTable = cdm$condition_occurrence,
                                            recordsPerPerson = c("mean", "sd"),
                                            inObservation = TRUE,
                                            standardConcept = TRUE,
                                            sourceVocabulary = TRUE,
                                            domainId = TRUE,
                                            typeConcept = TRUE)
tableClinicalRecords(summarisedResult)
}
}
