% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObservationPeriod.R
\name{tableObservationPeriod}
\alias{tableObservationPeriod}
\title{Create a visual table from a summariseObservationPeriod() result.}
\usage{
tableObservationPeriod(result, type = "gt")
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of formatting output table. See \code{visOmopResults::tableType()} for allowed options. Default is \code{"gt"}.}
}
\value{
A formatted table object with the summarised data.
}
\description{
Create a visual table from a summariseObservationPeriod() result.
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summariseObservationPeriod(observationPeriod = cdm$observation_period)

tableObservationPeriod(result = result)

PatientProfiles::mockDisconnect(cdm = cdm)
}
}
