% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOmopTable.R
\name{tableOmopTable}
\alias{tableOmopTable}
\title{Create a gt table from a summarised omop_table.}
\usage{
tableOmopTable(summarisedOmopTable)
}
\arguments{
\item{summarisedOmopTable}{A summarised_result object with the output from summariseOmopTable().}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
