% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePerson.R
\name{tablePerson}
\alias{tablePerson}
\title{Visualise the results of \code{summarisePerson()} into a table}
\usage{
tablePerson(result, type = "gt")
}
\arguments{
\item{result}{A summarised_result object created by \code{summarisePerson()}.}

\item{type}{One of the supported visualisation formats (see
\code{visOmopResults::tableType()}).}
}
\value{
A table visualisation.
}
\description{
Visualise the results of \code{summarisePerson()} into a table
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summarisePerson(cdm = cdm)

tablePerson(result = result)
}

}
