% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablePopulationCharacteristics.R
\name{tablePopulationCharacteristics}
\alias{tablePopulationCharacteristics}
\title{Create a gt table from a summarised population characteristics table.}
\usage{
tablePopulationCharacteristics(result, type = "gt")
}
\arguments{
\item{result}{Output from summarisePopulationCharacteristics().}

\item{type}{Type of table.}
}
\value{
A visual table.
}
\description{
Create a gt table from a summarised population characteristics table.
}
\examples{
\donttest{

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedPopulation <- summarisePopulationCharacteristics(
  cdm = cdm,
  studyPeriod = c("2010-01-01",NA),
  sex = TRUE,
  ageGroup = list("<=60" = c(0, 60), ">60" = c(61, Inf))
)

# Create a visual table
tablePopulationCharacteristics(summarisedPopulation)

# delete mock data
PatientProfiles::mockDisconnect(cdm = cdm)
}
}
