% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableQuality.R
\name{tableQuality}
\alias{tableQuality}
\title{Create a visual table from a summariseTableQuality() result.}
\usage{
tableQuality(result, type = "gt")
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of formatting output table. See \code{visOmopResults::tableType()} for allowed options. Default is \code{"gt"}.}
}
\value{
A formatted table object with the summarised data.
}
\description{
Create a visual table from a summariseTableQuality() result.
}
\examples{
\donttest{
cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summariseTableQuality(cdm, omopTableName = "drug_exposure")

tableQuality(result = result)

PatientProfiles::mockDisconnect(cdm = cdm)
}
}
