% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOmopResults.R
\name{visOmopResults.tidyColumns}
\alias{visOmopResults.tidyColumns}
\title{Get the tidy columns of a dataset}
\usage{
visOmopResults.tidyColumns(result)
}
\arguments{
\item{result}{A summarised_result object.}
}
\value{
A caharacter vector with the tidy columns.
}
\description{
Get the tidy columns of a dataset
}
\examples{
\donttest{
cdm <- mockOmopSketch()
result <- summariseOmopSnapshot(cdm)
visOmopResults.tidyColumns(result)
}

}
