% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockOmopSketch.R
\name{mockOmopSketch}
\alias{mockOmopSketch}
\title{Creates a mock database to test OmopSketch package.}
\usage{
mockOmopSketch(
  numberIndividuals = 100,
  con = lifecycle::deprecated(),
  writeSchema = lifecycle::deprecated(),
  seed = lifecycle::deprecated()
)
}
\arguments{
\item{numberIndividuals}{Number of individuals to create in the cdm
reference object.}

\item{con}{deprecated.}

\item{writeSchema}{deprecated.}

\item{seed}{deprecated.}
}
\value{
A mock cdm_reference object.
}
\description{
Creates a mock database to test OmopSketch package.
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)
cdm

# to insert into a duck db connection
library(duckdb)
library(CDMConnector)

con <- dbConnect(drv = duckdb())
to <- dbSource(con = con, writeSchema = "main")
cdm <- insertCdmTo(cdm = cdm, to = to)

cdm
}

}
