% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockOmopSketch.R
\name{mockOmopSketch}
\alias{mockOmopSketch}
\title{Creates a mock database to test OmopSketch package.}
\usage{
mockOmopSketch(con = NULL, writeSchema = NULL, numberIndividuals = 100)
}
\arguments{
\item{con}{A DBI connection to create the cdm mock object. By default, the
connection would be a 'duckdb' one.}

\item{writeSchema}{Name of an schema of the DBI connection with writing
permissions.}

\item{numberIndividuals}{Number of individuals to create in the cdm
reference object.}
}
\value{
A mock cdm_reference object.
}
\description{
Creates a mock database to test OmopSketch package.
}
\examples{
mockOmopSketch(numberIndividuals = 1000)

}
