% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockOmopSketch.R
\name{mockOmopSketch}
\alias{mockOmopSketch}
\title{Creates a mock database to test OmopSketch package}
\usage{
mockOmopSketch(
  numberIndividuals = 100,
  con = lifecycle::deprecated(),
  writeSchema = lifecycle::deprecated(),
  seed = lifecycle::deprecated()
)
}
\arguments{
\item{numberIndividuals}{Number of individuals to create in the cdm
reference object.}

\item{con}{deprecated.}

\item{writeSchema}{deprecated.}

\item{seed}{deprecated.}
}
\value{
A mock cdm_reference object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
