% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopSketch-package.R
\docType{package}
\name{OmopSketch-package}
\alias{OmopSketch}
\alias{OmopSketch-package}
\title{OmopSketch: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopSketch/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
