% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopSketch-package.R
\docType{package}
\name{OmopSketch-package}
\alias{OmopSketch}
\alias{OmopSketch-package}
\title{OmopSketch: Characterise Tables of an OMOP Common Data Model Instance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Summarises key information in data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model. Assess suitability to perform specific epidemiological studies and explore the different domains to obtain feasibility counts and trends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://OHDSI.github.io/OmopSketch/}
  \item Report bugs at \url{https://github.com/OHDSI/OmopSketch/issues}
}

}
\author{
\strong{Maintainer}: Cecilia Campanile \email{cecilia.campanile@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0007-6629-4661}{ORCID})

Authors:
\itemize{
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk} (\href{https://orcid.org/0000-0002-8462-8668}{ORCID})
  \item Elin Rowlands \email{elin.rowlands@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0005-5166-0417}{ORCID})
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
