% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConceptCounts.R
\name{plotConceptCounts}
\alias{plotConceptCounts}
\title{Title}
\usage{
plotConceptCounts(result, facet = NULL, colour = NULL)
}
\arguments{
\item{result}{A summarised_result object (output of summariseInObservation).}

\item{facet}{Columns to face by. Formula format can be provided. See possible
columns to face by with: \code{visOmopResults::tidyColumns()}.}

\item{colour}{Columns to colour by. See possible columns to colour by with:
\code{visOmopResults::tidyColumns()}.}
}
\value{
A ggplot showing the concept counts
}
\description{
Title
}
\examples{
\dontrun{
library(OmopSketch)
library(dplyr)

con <- DBI::dbConnect(duckdb::duckdb(),
                      dbdir = CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(con,
                                  cdm_schem = "main",
                                  write_schema = "main")

results <- summariseConceptCounts(cdm,
         conceptId = list(poliovirus_vaccine = c(40213160)))

plotConceptCounts(result)

PatientProfiles::mockDisconnect(cdm)
}
}
