% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInObservation.R
\name{plotInObservation}
\alias{plotInObservation}
\title{Create a gt table from a summarised omop_table.}
\usage{
plotInObservation(summarisedInObservation, facet = NULL)
}
\arguments{
\item{summarisedInObservation}{Output from summariseInObservation().}

\item{facet}{columns in data to facet. If the facet position wants to be specified, use the formula class for the input
(e.g., strata_level ~ group_level + cdm_name). Variables before "~" will be facet by on horizontal axis, whereas those after "~" on vertical axis.
Only the following columns are allowed to be facet by: "cdm_name", "group_level", "strata_level".}
}
\value{
A ggplot showing the table counts
}
\description{
Create a gt table from a summarised omop_table.
}
