% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObservationPeriod.R
\name{plotObservationPeriod}
\alias{plotObservationPeriod}
\title{Create a gt table from a summarised omop_table.}
\usage{
plotObservationPeriod(summarisedObservationPeriod)
}
\arguments{
\item{summarisedObservationPeriod}{A summarised_result object with the output from summariseObservationPeriod().}
}
\value{
A ggplot showing the table counts
}
\description{
Create a gt table from a summarised omop_table.
}
