% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObservationPeriod.R
\name{plotObservationPeriod}
\alias{plotObservationPeriod}
\title{Create a plot from the output of summariseObservationPeriod().}
\usage{
plotObservationPeriod(
  result,
  variableName = "Number subjects",
  plotType = "barplot",
  facet = NULL,
  colour = NULL,
  style = NULL
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{variableName}{The variable to plot it can be: "Number subjects",
"Records per person", "Duration in days" or "Days to next observation period".}

\item{plotType}{The plot type, it can be: "barplot", "boxplot" or
"densityplot".}

\item{facet}{Columns to colour by. See possible columns to colour by with:
\code{visOmopResults::tidyColumns()}.}

\item{colour}{Columns to colour by. See possible columns to colour by with:
\code{visOmopResults::tidyColumns()}.}

\item{style}{Visual theme to apply. Character, or \code{NULL}. If a character,
this may be either the name of a built-in style (see \code{plotStyle()}), or a
path to a \code{.yml} file that defines a custom style. If \code{NULL}, the function
will use the explicit default style, unless a global style option is set (see
\code{setGlobalPlotOptions()}), or a \verb{⁠_brand.yml}⁠ file is present (in that order).
Refer to the package vignette on styles to learn more.}
}
\value{
A ggplot2 object.
}
\description{
Create a plot from the output of summariseObservationPeriod().
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summariseObservationPeriod(observationPeriod = cdm$observation_period)

plotObservationPeriod(
  result = result,
  variableName = "Duration in days",
  plotType = "boxplot"
)

PatientProfiles::mockDisconnect(cdm)
}
}
