% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecordCount.R
\name{plotRecordCount}
\alias{plotRecordCount}
\title{Create a gt table from a summarised omop_table.}
\usage{
plotRecordCount(summarisedRecordCount)
}
\arguments{
\item{summarisedRecordCount}{A summarised_result object with the output from summariseTableCounts().}
}
\value{
A ggplot showing the table counts
}
\description{
Create a gt table from a summarised omop_table.
}
