% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTableCounts.R
\name{plotTableCounts}
\alias{plotTableCounts}
\title{Create a gt table from a summarised omop_table.}
\usage{
plotTableCounts(summarisedTableCounts)
}
\arguments{
\item{summarisedTableCounts}{A summarised_result object with the output from summariseTableCounts().}
}
\value{
A ggplot showing the table counts
}
\description{
Create a gt table from a summarised omop_table.
}
