% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{style-plot}
\alias{style-plot}
\title{Helper for consistent documentation of \code{style}.}
\arguments{
\item{style}{Which style to apply to the plot, options are: "default", "darwin"
and NULL (default ggplot style). Customised styles can be achieved by modifying the returned ggplot object.}
}
\description{
Helper for consistent documentation of \code{style}.
}
\keyword{internal}
