% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseObservationPeriod.R
\name{summariseObservationPeriod}
\alias{summariseObservationPeriod}
\title{Summarise the observation period table getting some overall statistics in a
summarised_result object.}
\usage{
summariseObservationPeriod(observationPeriod, density = FALSE)
}
\arguments{
\item{observationPeriod}{observation_period omop table}

\item{density}{Whether to export density data for time between observation
periods.}
}
\value{
A summarised_result object with the summarised data.
}
\description{
Summarise the observation period table getting some overall statistics in a
summarised_result object.
}
