% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOmopSnapshot.R
\name{summariseOmopSnapshot}
\alias{summariseOmopSnapshot}
\title{Summarise OMOP database and create snapshot}
\usage{
summariseOmopSnapshot(cdm)
}
\arguments{
\item{cdm}{A cdm reference object}
}
\value{
A summarised result object
}
\description{
Summarise OMOP database and create snapshot
}
\examples{
\donttest{
library(dplyr)
library(OmopSketch)

# Connect to Eunomia database

cdm <- mockOmopSketch(numberIndividuals = 1000)

# Run OMOP database snapshot

snapshot <- summariseOmopSnapshot(cdm)
}
}
