% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOmopSnapshot.R
\name{summariseOmopSnapshot}
\alias{summariseOmopSnapshot}
\title{Summarise a cdm_reference object creating a snapshot with the metadata of the
cdm_reference object.}
\usage{
summariseOmopSnapshot(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A summarised_result object.
}
\description{
Summarise a cdm_reference object creating a snapshot with the metadata of the
cdm_reference object.
}
\examples{
\donttest{
cdm <- mockOmopSketch(numberIndividuals = 10)

summariseOmopSnapshot(cdm)
}
}
