% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePerson.R
\name{summarisePerson}
\alias{summarisePerson}
\title{Summarise the person table}
\usage{
summarisePerson(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A summarised_result object with the summary of the person table.
}
\description{
Summarise the person table
}
\examples{
\donttest{
library(dplyr, warn.conflicts = FALSE)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summarisePerson(cdm = cdm)

result |>
  glimpse()

PatientProfiles::mockDisconnect(cdm)
}

}
