% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePopulationCharacteristics.R
\name{summarisePopulationCharacteristics}
\alias{summarisePopulationCharacteristics}
\title{Summarise the characteristics of the base population of a cdm object.}
\usage{
summarisePopulationCharacteristics(
  cdm,
  studyPeriod = c(NA, NA),
  sex = FALSE,
  ageGroup = NULL
)
}
\arguments{
\item{cdm}{A cdm object.}

\item{studyPeriod}{Dates to trim the observation period. If NA, start_observation_period and/or end_observation_period are used.}

\item{sex}{Boolean variable. Whether to stratify the results by sex.}

\item{ageGroup}{List of age groups to stratify by at index date. Set to NULL if no stratification is needed.}
}
\value{
A summarised_result object.
}
\description{
Summarise the characteristics of the base population of a cdm object.
}
\examples{
\donttest{
library(dplyr)
library(CDMConnector)
library(DBI)
library(duckdb)
library(OmopSketch)

# Connect to Eunomia database
if (Sys.getenv("EUNOMIA_DATA_FOLDER") == "") Sys.setenv("EUNOMIA_DATA_FOLDER" = tempdir())
if (!dir.exists(Sys.getenv("EUNOMIA_DATA_FOLDER"))) dir.create(Sys.getenv("EUNOMIA_DATA_FOLDER"))
if (!eunomia_is_available()) downloadEunomiaData()
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdmFromCon(
con = con, cdmSchema = "main", writeSchema = "main"
)

# Run summarise clinical tables
summarisedPopulation <- summarisePopulationCharacteristics(cdm = cdm,
                                                          studyPeriod = c("2010-01-01",NA),
                                                          sex = TRUE,
                                                          ageGroup = NULL
                                                          )
summarisedPopulation |> print()
PatientProfiles::mockDisconnect(cdm = cdm)
}
}
