% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseRecordCount.R
\name{summariseRecordCount}
\alias{summariseRecordCount}
\title{Create a gt table from a summarised omop_table.}
\usage{
summariseRecordCount(
  omopTable,
  unit = "year",
  unitInterval = 1,
  ageGroup = NULL
)
}
\arguments{
\item{omopTable}{A summarised_result object with the output from summariseOmopTable().}

\item{unit}{Whether to stratify by "year" or by "month"}

\item{unitInterval}{Number of years or months to stratify with}

\item{ageGroup}{A list of age groups to stratify results by.}
}
\value{
A summarised_result object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
