% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseRecordCount.R
\name{summariseRecordCount}
\alias{summariseRecordCount}
\title{Create a summarise result object to summarise record counts of an omop_table using a specific time interval. Only records that fall within the observation period are counted.}
\usage{
summariseRecordCount(
  cdm,
  omopTableName,
  unit = "year",
  unitInterval = 1,
  ageGroup = NULL,
  sex = FALSE
)
}
\arguments{
\item{cdm}{A cdm object.}

\item{omopTableName}{A character vector of omop tables from the cdm.}

\item{unit}{Whether to stratify by "year" or by "month".}

\item{unitInterval}{An integer. Number of years or months to include within the same interval.}

\item{ageGroup}{A list of age groups to stratify results by.}

\item{sex}{Boolean variable. Whether to stratify by sex (TRUE) or not (FALSE).}
}
\value{
A summarised_result object.
}
\description{
Create a summarise result object to summarise record counts of an omop_table using a specific time interval. Only records that fall within the observation period are counted.
}
\examples{
\donttest{

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedResult <- summariseRecordCount(cdm = cdm,
                                         omopTableName = "condition_occurrence",
                                         unit = "year",
                                         unitInterval = 10,
                                         ageGroup = list("<=20" = c(0,20), ">20" = c(21, Inf)),
                                         sex = TRUE)
summarisedResult |> print()
}
}
