% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a gt table from a summarised omop_table.}
\usage{
tableClinicalRecords(summarisedClinicalRecords)
}
\arguments{
\item{summarisedClinicalRecords}{A summarised_result object with the output from summariseClinicalRecords().}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
