% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a gt table from a summarised omop_table.}
\usage{
tableClinicalRecords(summarisedClinicalRecords)
}
\arguments{
\item{summarisedClinicalRecords}{Output from summariseClinicalRecords().}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
\examples{
\donttest{
library(dplyr)
library(CDMConnector)
library(DBI)
library(duckdb)
library(OmopSketch)

# Connect to Eunomia database
if (Sys.getenv("EUNOMIA_DATA_FOLDER") == "") Sys.setenv("EUNOMIA_DATA_FOLDER" = tempdir())
if (!dir.exists(Sys.getenv("EUNOMIA_DATA_FOLDER"))) dir.create(Sys.getenv("EUNOMIA_DATA_FOLDER"))
if (!eunomia_is_available()) downloadEunomiaData()
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdmFromCon(
con = con, cdmSchema = "main", writeSchema = "main"
)

# Run summarise clinical tables
summarisedResult <- summariseClinicalRecords(cdm = cdm,
                                            omopTableName = "condition_occurrence",
                                            recordsPerPerson = c("mean", "sd"),
                                            inObservation = TRUE,
                                            standardConcept = TRUE,
                                            sourceVocabulary = TRUE,
                                            domainId = TRUE,
                                            typeConcept = TRUE)
tableClinicalRecords(summarisedResult)
}
}
