% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a visual table from a summariseClinicalRecord() output.}
\usage{
tableClinicalRecords(result, type = NULL, style = NULL)
}
\arguments{
\item{result}{Output from summariseClinicalRecords().}

\item{type}{Character string specifying the desired output table format. See
\code{tableType()} for supported table types. If \code{type = NULL}, global options
(set via \code{setGlobalTableOptions()}) will be used if available; otherwise, a
default 'gt' table is created.}

\item{style}{Defines the visual formatting of the table. This argument can be
provided in one of the following ways:
\enumerate{
\item \strong{Pre-defined style}: Use the name of a built-in style (e.g., "darwin").
See \code{tableStyle()} for available options.
\item \strong{YAML file path}: Provide the path to an existing \code{.yml} file defining
a new style.
\item \strong{List of custome R code}: Supply a block of custom R code or a named
list describing styles for each table section. This code must be specific to
the selected table type.
}

If \code{style = NULL}, the function will use global options (see
\code{setGlobalTableOptions()}) or an existing \verb{⁠_brand.yml}⁠ file (if found);
otherwise, the default style is applied. For more details, see the Styles
vignette on the package website.}
}
\value{
A formatted table object with the summarised data.
}
\description{
Create a visual table from a summariseClinicalRecord() output.
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch()

summarisedResult <- summariseClinicalRecords(
  cdm = cdm,
  omopTableName = c("condition_occurrence", "drug_exposure"),
  recordsPerPerson = c("mean", "sd"),
  inObservation = TRUE,
  standardConcept = TRUE,
  sourceVocabulary = TRUE,
  domainId = TRUE,
  typeConcept = TRUE
)

summarisedResult |>
  suppress(minCellCount = 5) |>
  tableClinicalRecords()

PatientProfiles::mockDisconnect(cdm)
}
}
