% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a visual table from a summarise_clinical_record result.}
\usage{
tableClinicalRecords(result, type = "gt")
}
\arguments{
\item{result}{Output from summariseClinicalRecords().}

\item{type}{Type of formatting output table, either "gt" or "flextable".}
}
\value{
A gt or flextable object with the summarised data.
}
\description{
Create a visual table from a summarise_clinical_record result.
}
\examples{
\donttest{

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedResult <- summariseClinicalRecords(
  cdm = cdm,
  omopTableName = "condition_occurrence",
  recordsPerPerson = c("mean", "sd"),
  inObservation = TRUE,
  standardConcept = TRUE,
  sourceVocabulary = TRUE,
  domainId = TRUE,
  typeConcept = TRUE)

tableClinicalRecords(summarisedResult)

PatientProfiles::mockDisconnect(cdm)
}
}
