% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClinicalRecords.R
\name{tableClinicalRecords}
\alias{tableClinicalRecords}
\title{Create a gt table from a summarised omop_table.}
\usage{
tableClinicalRecords(result, type = "gt")
}
\arguments{
\item{result}{Output from summariseClinicalRecords().}

\item{type}{type of table.}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop_table.
}
\examples{
\donttest{

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedResult <- summariseClinicalRecords(
  cdm = cdm,
  omopTableName = "condition_occurrence",
  recordsPerPerson = c("mean", "sd"),
  inObservation = TRUE,
  standardConcept = TRUE,
  sourceVocabulary = TRUE,
  domainId = TRUE,
  typeConcept = TRUE)

tableClinicalRecords(summarisedResult)

PatientProfiles::mockDisconnect(cdm)
}
}
