% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObservationPeriod.R
\name{tableObservationPeriod}
\alias{tableObservationPeriod}
\title{Create a visual table from a summariseObservationPeriod() result.}
\usage{
tableObservationPeriod(result, type = "gt", style = NULL)
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of formatting output table. See \code{visOmopResults::tableType()} for allowed options. Default is \code{"gt"}.}

\item{style}{Defines the visual formatting of the table. This argument can be
provided in one of the following ways:
\enumerate{
\item \strong{Pre-defined style}: Use the name of a built-in style (e.g., "darwin").
See \code{tableStyle()} for available options.
\item \strong{YAML file path}: Provide the path to an existing \code{.yml} file defining
a new style.
\item \strong{List of custome R code}: Supply a block of custom R code or a named
list describing styles for each table section. This code must be specific to
the selected table type.
}

If \code{style = NULL}, the function will use global options (see
\code{setGlobalTableOptions()}) or an existing \verb{⁠_brand.yml}⁠ file (if found);
otherwise, the default style is applied. For more details, see the Styles
vignette on the package website.}
}
\value{
A formatted table object with the summarised data.
}
\description{
Create a visual table from a summariseObservationPeriod() result.
}
\examples{
\donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summariseObservationPeriod(observationPeriod = cdm$observation_period)

tableObservationPeriod(result = result)

CDMConnector::cdmDisconnect(cdm = cdm)
}
}
