% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObservationPeriod.R
\name{tableObservationPeriod}
\alias{tableObservationPeriod}
\title{Create a table from the output of summariseObservationPeriod().}
\usage{
tableObservationPeriod(result, type = "gt")
}
\arguments{
\item{result}{A summarised result object.}

\item{type}{Type of table.}
}
\value{
A gt or flextable table.
}
\description{
Create a table from the output of summariseObservationPeriod().
}
