% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObservationPeriod.R
\name{tableObservationPeriod}
\alias{tableObservationPeriod}
\title{Create a visual table from a summariseObservationPeriod() result.}
\usage{
tableObservationPeriod(result, type = "gt")
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of formatting output table, either "gt" or "flextable".}
}
\value{
A gt or flextable object with the summarised data.
}
\description{
Create a visual table from a summariseObservationPeriod() result.
}
\examples{
\donttest{
cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summariseObservationPeriod(cdm$observation_period)

tableObservationPeriod(result)

PatientProfiles::mockDisconnect(cdm)
}
}
