% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOmopSnapshot.R
\name{tableOmopSnapshot}
\alias{tableOmopSnapshot}
\title{Create a gt table from a summarised omop snapshot}
\usage{
tableOmopSnapshot(result, type = "gt")
}
\arguments{
\item{result}{Output from summariseOmopSnapshot().}

\item{type}{Type of table.}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised omop snapshot
}
\examples{
\donttest{
cdm <- mockOmopSketch(numberIndividuals = 1000)

cdm |>
  summariseOmopSnapshot() |>
  tableOmopSnapshot()

PatientProfiles::mockDisconnect(cdm)
}
}
