% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOmopSnapshot.R
\name{tableOmopSnapshot}
\alias{tableOmopSnapshot}
\title{Create a visual table from a summarise_omop_snapshot result.}
\usage{
tableOmopSnapshot(result, type = "gt")
}
\arguments{
\item{result}{Output from summariseOmopSnapshot().}

\item{type}{Type of formatting output table, either "gt" or "flextable".}
}
\value{
A gt or flextable object with the summarised data.
}
\description{
Create a visual table from a summarise_omop_snapshot result.
}
\examples{
cdm <- mockOmopSketch(numberIndividuals = 10)

result <- summariseOmopSnapshot(cdm)

result |>
  tableOmopSnapshot()

PatientProfiles::mockDisconnect(cdm)
}
