% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOmopTable.R
\name{tableOmopTable}
\alias{tableOmopTable}
\title{Summarise an omop_table from a cdm_reference object. You will obtain
information related to the number of records, number of subjects, whether the
records are in observation, number of present domains and number of present
concepts.}
\usage{
tableOmopTable(summarisedOmopTable)
}
\arguments{
\item{summarisedOmopTable}{A summarised_result object with the output from summariseOmopTable().}
}
\value{
A gt object with the summarised data.
}
\description{
Summarise an omop_table from a cdm_reference object. You will obtain
information related to the number of records, number of subjects, whether the
records are in observation, number of present domains and number of present
concepts.
}
