% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePerson.R, R/tablePerson.R
\name{tablePerson}
\alias{tablePerson}
\title{Visualise the results of \code{summarisePerson()} into a table}
\usage{
tablePerson(result, style = "default", type = "gt")

tablePerson(result, style = "default", type = "gt")
}
\arguments{
\item{result}{A summarised_result object.}

\item{style}{A character string or custom R code to define the visual
formatting of the table.}

\item{type}{The desired format of the output table. See
\code{visOmopResults::tableType()} for allowed options.}
}
\value{
A table visualisation.

A visualisation of the data summarising the person table.
}
\description{
Visualise the results of \code{summarisePerson()} into a table

Visualise the output of \code{summarisePerson()}
}
\examples{
\donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

result <- summarisePerson(cdm = cdm)

tablePerson(result = result)
}

\donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

result <- summarisePerson(cdm = cdm)

tablePerson(result = result)
}

}
