% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablePopulationCharacteristics.R
\name{tablePopulationCharacteristics}
\alias{tablePopulationCharacteristics}
\title{Create a gt table from a summarised population characteristics table.}
\usage{
tablePopulationCharacteristics(result)
}
\arguments{
\item{result}{Output from summarisePopulationCharacteristics().}
}
\value{
A gt object with the summarised data.
}
\description{
Create a gt table from a summarised population characteristics table.
}
\examples{
\donttest{
library(dplyr)
library(OmopSketch)

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedPopulation <- summarisePopulationCharacteristics(
                       cdm = cdm,
                       studyPeriod = c("2010-01-01",NA),
                       sex = TRUE,
                       ageGroup = list("<=60" = c(0, 60), ">60" = c(61, Inf))
                       )
tablePopulationCharacteristics(summarisedPopulation)
PatientProfiles::mockDisconnect(cdm = cdm)
}
}
